+++ E2/parameters.py
@@ -5,13 +5,12 @@ print("DEBUG", sys.argv)
 days = ['L', 'M', 'W', 'J', 'V']
 salas = ['A', 'B', 'C', 'D', 'E', 'F', 'G']
 
-CAMBIAR_PERSONAS = 0
-CAMBIAR_SALAS = 0
-CAMBIAR_MODULOS = 0
-CAMBIAR_REUNIONES = 0
+CAMBIAR_PERSONAS = 1
+CAMBIAR_SALAS = 1
+CAMBIAR_MODULOS = 1
+CAMBIAR_REUNIONES = 1
 
 if len(sys.argv) > 1:
-    print("Holaaaa")
     CANTIDAD_PERSONAS = int(sys.argv[1])
     CANTIDAD_ESPACIOS_FISICOS = int(sys.argv[2])
     CANTIDAD_MODULOS = int(sys.argv[3])
@@ -29,10 +28,10 @@ else:
     CANTIDAD_MODULOS = 9
     CANTIDAD_REUNIONES = 5
 
-    MIN_A = 50
-    MAX_A = 100
+    MIN_A = 0
+    MAX_A = 25
 
     MIN_L = 0
-    MAX_L = 25
+    MAX_L = 5
 
 M_GRANDE = 999999
